Update README.md
This is the repository for DLTAfrica website.

This repository accepts contributions via GitHub pull requests. This document outlines some contributing guidelines, contact points, and other resources to make it easier to contribute to this repository.

If you've run into a problem or have a suggestion, browse the open issues before opening a new one. We use the following label system to organize our  issues.